Remove duplicated code
There was a rather large piece of code that was mostly the
same between VersionedQuerySupport and TemporalQuerySupport.
This change moves that into Utils.

Also a few issues reported by IntelliJ's Inspections have
been taken care of

+++ grooves-diagrams/src/main/groovy/com/github/rahulsom/grooves/asciidoctor/Event.groovy
@@ -40,7 +40,7 @@ class Event {
                 x += 10
             }
             this.x = x
-            this.y = y
+            this.y = (int) y
 
             if (type == EventType.Revert) {
                 def revertedId = description.split(' ')[-1]
